requirements.txt
Django==5.0
djangorestframework
requests
gunicorn
celery
django-filter